requirements.txt
torch==2.0.1
transformers==4.30.2
Pillow==9.5.0
pandas==2.0.3
scikit-learn==1.2.2
matplotlib==3.7.1
seaborn==0.12.2
